requirements.txt
PySide6>=6.5
python-can>=4.0
pyusb>=1.0
